Restructured API filestructure
